refactor(ui)!: finish the root realignment — node clients into components/, helpers into lib/
text-link and image-preview move next to their node folders
(components/link/, components/image/), and the entry.* renderers leave
the stray entries/ directory for components/entries/ — components/ now
holds everything that renders wire nodes. The loose style and hook
helpers (pill, control, use-collapsible-state, column-sizing,
use-column-resizing) move into lib/. The package root is left to runtime
infrastructure and the generic client primitives.

BREAKING CHANGE: deep imports move with the files —
@lattice-php/ui/control, /column-sizing, /use-column-resizing, and
/use-collapsible-state are now under @lattice-php/ui/lib/*, and
@lattice-php/ui/image-preview is @lattice-php/ui/components/image/image-preview.
Root-index imports are unaffected.
