+++ packages/ui/resources/js/components/collapsible/collapsible.tsx
@@ -3,7 +3,7 @@ import type { ComponentProps, KeyboardEvent, MouseEvent, ReactNode } from "react
 import { Icon } from "../../icons";
 import { InfoTooltip } from "../../info-tooltip";
 import { cn } from "../../lib/utils";
-import { useCollapsibleState } from "../../use-collapsible-state";
+import { useCollapsibleState } from "../../lib/use-collapsible-state";
 
 export type CollapsibleTriggerProps = Omit<
   ComponentProps<"div">,
