+++ packages/ui/resources/js/components/image/image-preview.browser.test.tsx
