+++ packages/ui/resources/js/components/section/section.tsx
@@ -2,7 +2,7 @@ import { type ComponentProps, type MouseEvent, type ReactNode } from "react";
 import { Icon } from "../../icons";
 import { InfoTooltip } from "../../info-tooltip";
 import { cn } from "../../lib/utils";
-import { useCollapsibleState } from "../../use-collapsible-state";
+import { useCollapsibleState } from "../../lib/use-collapsible-state";
 
 export type SectionToggleProps = Omit<
   ComponentProps<"button">,
