+++ packages/ui/resources/js/index.ts
@@ -4,7 +4,7 @@ export * from "./components/badge/badge";
 export * from "./components/button/button";
 export * from "./checkbox";
 export * from "./color-picker";
-export * from "./column-sizing";
+export * from "./lib/column-sizing";
 export * from "./components/avatar/avatar";
 export * from "./components/collapsible/collapsible";
 export * from "./components/description-list/description-list";
@@ -22,7 +22,7 @@ export * from "./components/text/text";
 export * from "./components/tooltip/tooltip";
 export * from "./combobox";
 export * from "./confirm-dialog";
-export * from "./control";
+export * from "./lib/control";
 export {
   ActionTrigger,
   ActionTriggerProvider,
@@ -57,7 +57,7 @@ export * from "./icon-button";
 export * from "./icons/index";
 export * from "./i18n/index";
 export * from "./i18n/translatable";
-export * from "./image-preview";
+export * from "./components/image/image-preview";
 export * from "./info-tooltip";
 export * from "./input";
 export { default as InputError } from "./input-error";
@@ -71,10 +71,10 @@ export * from "./popover";
 export * from "./skeleton";
 export * from "./spinner";
 export * from "./textarea";
-export * from "./text-link";
+export * from "./components/link/text-link";
 export * from "./types";
-export * from "./use-collapsible-state";
-export * from "./use-column-resizing";
+export * from "./lib/use-collapsible-state";
+export * from "./lib/use-column-resizing";
 export * from "./format/temporal";
 export * from "./format/number";
 export * from "./format/numeric";
